+++ src/main/java/com/example/demo/dto/Co2ResponseDto.java
@@ -3,6 +3,9 @@ package com.example.demo.dto;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 
 @Getter
@@ -11,7 +14,7 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public class Co2ResponseDto {
     private String status;
-    private double reduction;
+    private BigDecimal reduction;
     private String message;
     private Long recordId;
 }
